Fix problem with no preview when affe-grep from a subdir of a project
Assume that the current buffer is visiting ~/proj/a/b/c/d/e.txt,
where ~/proj is the root directory of the project.

Invoke affe-grep and press preview-key, we can't preview the file.

The reason is as follows.

When consult opens a file for preview, a function returned from consult--temporary-files
searches it from default-directory.
Therefore, default-directory must be properly set before the call of consult--temporary-files.
Since consult--grep-state calls consult--temporary-files, consult--grep-state must be called
after default-directory is set.